requirements.txt
django == 2.2.20
djangorestframework == 3.12.4